Player Projectile implemented
Player projectile concept gives the possibility of a player to attack from range. Currently, this is possible for Mage and Archer.

Dev. summary: projectile is a weapon used for range attack. Collision of projectile with the monsters is implemented, the damage caused to the enemy and the animations along with the sound repetations are handled.